Git: Always create temp branches from the pre-sync HEAD commit.
Until now, the temp branches ("conflict branches") which get created upon
sync conflicts have used as their starting point the last synced commit
of the notebook/file in question. For notebooks that don't change very
often, this can be far back in the repo history.

When we checkout a new branch starting from an old commit, the whole
working directory is reset to that point in history. This often means
that other notebooks are reverted to older versions, leading to
confusing results during subsequent syncing, and often the apparent loss
of work (until the branches have been merged).

The desired behavior in the event of a merge conflict is instead (in my
opinion) that Orgzly's view of the repo be preserved, only remote
changes do not become available in Orgzly until the merge conflict has
been resolved.

In order to achieve this, we use a special "marker" branch, which we
move to the current head before each sync occasion. This branch then
later provides us with a good starting point for branching off in case
of a conflict.