Sort namesakes before syncing them,
ensuring that notebooks in Git repos with conflicting changes get synced first,
so that any required "conflict branch" is created as early as possible.

Otherwise, we risk committing non-conflicting changes to the master branch --
changes which would later seemingly disappear when we checkout a conflict
branch with the pre-sync HEAD commit as its starting point.